// src/api/generated/model/userDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chargee Ampere
 * API v2
 * OpenAPI spec version: 2.0.0
 */
import type { UserDtoRole } from './userDtoRole';

export interface UserDto {
  /** The unique identifier of the user */
  uuid: string;
  /** The role of the user */
  role: UserDtoRole;
  /** The email of the user */
  email: string;
  /** The time this resource was created. */
  createdAt: string;
  /**
   * The time this resource was updated.
   * @nullable
   */
  updatedAt: string | null;
  /**
   * The time the user last logged in.
   * @nullable
   */
  loginAt: string | null;
  /**
   * The time this resource was deleted.
   * @nullable
   */
  deletedAt: string | null;
}